requirements.txt
asgiref==3.8.1
Django==5.2.4
django-admin-thumbnails==0.2.9
pillow==11.3.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
